fix:weekly attendace view
