+++ apps/front-end/src/pages/front-end/Camp/Attendace/CampAttendance.js
@@ -1,18 +1,23 @@
 import React from "react";
-import { Avatar, Box, HStack, Pressable, Stack, VStack } from "native-base";
+import { Box, HStack, VStack } from "native-base";
 import {
   Layout,
   FrontEndTypo,
   AdminTypo,
-  ImageView,
   IconByName,
   campService,
   Camera,
   Loading,
   uploadRegistryService,
+  GeoLocation,
+  UserCard,
 } from "@shiksha/common-lib";
 import { useNavigate, useParams } from "react-router-dom";
 import { useTranslation } from "react-i18next";
+import Chip from "component/Chip";
+
+const PRESENT = "present";
+const ABSENT = "absent";
 
 // App
 export default function ConsentForm() {
@@ -21,24 +26,31 @@ export default function ConsentForm() {
   const navigate = useNavigate();
   const { t } = useTranslation();
   const [groupUsers, setGroupUsers] = React.useState();
-  const [consents, setConsents] = React.useState();
-  const [cameraModal, setCameraModal] = React.useState(false);
   const [cameraUrl, setCameraUrl] = React.useState();
   const [userData, setUserData] = React.useState({});
   const [error, setError] = React.useState("");
-  const [cameraFile, setcameraFile] = React.useState();
+  const [cameraFile, setCameraFile] = React.useState();
+  const [data, setData] = React.useState({});
 
   React.useEffect(async () => {
+    await getData();
+  }, [id, !userData]);
+
+  const getData = async () => {
     const result = await campService.getCampDetails({ id });
-    const campConsent = await campService.CampAttendance({ id });
-    if (Object.keys(campConsent?.data).length === 0) {
-      setConsents([]);
-    } else {
-      setConsents(campConsent?.data);
+    const resultAttendance = await campService.CampAttendance({ id });
+    let attendances = [];
+    if (resultAttendance?.data?.length > 0) {
+      attendances = resultAttendance?.data;
     }
-    setGroupUsers(result?.data?.group_users);
+    setGroupUsers(
+      result?.data?.group_users.map((item, index) => {
+        let attendance = attendances.find((e) => e?.user?.id === item.id);
+        return { ...item, index, attendance };
+      })
+    );
     setLoading(false);
-  }, [id]);
+  };
 
   // update schema
 
@@ -48,39 +60,53 @@ export default function ConsentForm() {
 
   // Camera MOdule
 
-  const uploadAttendencePicture = async (e) => {
+  const uploadAttendence = async (user, status = PRESENT, finish = false) => {
     setError("");
-    if (cameraFile?.key) {
-      await campService.markCampAttendance({
-        context: "camp",
-        context_id: id,
-        user_id: userData?.id,
-        status: "present",
-        photo_1: cameraFile ? cameraFile?.key : "",
-      });
+    if (user?.attendance?.status) {
+      if (status === PRESENT || status === ABSENT) {
+        let payLoad = {
+          ...data,
+          id: user?.attendance?.id,
+          context_id: id,
+          user_id: user?.id,
+          status,
+        };
+        if (status === PRESENT) {
+          const photo_1 =
+            cameraFile?.data?.insert_documents?.returning?.[0]?.id;
+          payLoad = { ...payLoad, photo_1: `${photo_1}` };
+        }
+        await campService.updateCampAttendance(payLoad);
+        await getData();
+      }
     } else {
-      setError("Capture Picture First");
-    }
-    const coruntIndex = groupUsers.findIndex(
-      (item) => item?.id === userData?.id
-    );
-    if (groupUsers[coruntIndex + 1]) {
-      setCameraUrl();
-      setUserData({ ...groupUsers[coruntIndex + 1], index: coruntIndex + 1 });
-    }
-  };
+      const photo_1 = cameraFile?.data?.insert_documents?.returning?.[0]?.id;
+      if (photo_1) {
+        const payLoad = {
+          ...data,
+          context_id: id,
+          user_id: user?.id,
+          status: PRESENT,
+          photo_1: `${photo_1}`,
+        };
 
-  const updateUserData = async () => {
-    if (cameraFile?.key) {
-      await campService.markCampAttendance({
-        context: "camp",
-        context_id: id,
-        user_id: userData?.id,
-        status: "present",
-        photo_1: cameraFile ? cameraFile?.key : "",
-      });
+        await campService.markCampAttendance(payLoad);
+        await getData();
+      } else {
+        setError("Capture Picture First");
+      }
+    }
+
+    if (finish) {
+      setCameraUrl();
+      setCameraFile();
+      setUserData();
     } else {
-      setError("Capture Picture First");
+      const coruntIndex = groupUsers.findIndex((item) => item?.id === user?.id);
+      if (groupUsers[coruntIndex + 1]) {
+        setCameraUrl();
+        setUserData({ ...groupUsers[coruntIndex + 1], index: coruntIndex + 1 });
+      }
     }
   };
 
@@ -91,97 +117,88 @@ export default function ConsentForm() {
           <React.Suspense fallback={<Loading />}>
             <Camera
               headerComponent={
-                <VStack bg="black" width="94%" pl="4">
-                  {/* <AdminTypo.H6 color="white" bold>
-                    {t("MARK_ATTENDANCE_ORIENTATION")}
-                  </AdminTypo.H6>
-                  <HStack direction={["row", "row", "row"]}>
-                    <AdminTypo.H6 color="white" bold flex="0.3">
-                      {t("PRESENT")} :
-                      {users.filter((e) => e.status === "present").length}
-                    </AdminTypo.H6>
-                    <AdminTypo.H6 color="white" bold flex="0.3">
-                      {t("ABSENT")} :
-                      {users.filter((e) => e.status !== "present").length}
-                    </AdminTypo.H6>
-                  </HStack> */}
-                  <HStack direction={["row", "row", "row"]}>
+                <VStack bg="black" flex="1" py="2" px="4">
+                  <HStack
+                    space={2}
+                    divider={
+                      <AdminTypo.H6 color="white" bold>
+                        :
+                      </AdminTypo.H6>
+                    }
+                  >
+                    <AdminTypo.H6 color="white">{t("NAME")}</AdminTypo.H6>
                     <AdminTypo.H6 color="white">
-                      {t("CANDIDATES_NAME")}{" "}
-                      {
+                      {/* ${userData?.index + 1}) */}
+                      {`${[
                         userData?.program_beneficiaries[0]
-                          ?.enrollment_first_name
-                      }
+                          ?.enrollment_first_name,
+                        userData?.program_beneficiaries[0]
+                          ?.enrollment_middle_name,
+                        userData?.program_beneficiaries[0]
+                          ?.enrollment_last_name,
+                      ]
+                        .filter((e) => e)
+                        .join(" ")}`}
                     </AdminTypo.H6>
                   </HStack>
-                  <HStack>
-                    <AdminTypo.H6
-                      color="white"
-                      direction={["row", "row", "row"]}
-                    >
-                      {t("CANDIDATES")} -{" "}
+                  {/* <HStack
+                    space={2}
+                    divider={
+                      <AdminTypo.H6 color="white" bold>
+                        :
+                      </AdminTypo.H6>
+                    }
+                  >
+                    <AdminTypo.H6 color="white">{t("CANDIDATES")}</AdminTypo.H6>
+                    <AdminTypo.H6 color="white">
                       {groupUsers?.length ? groupUsers?.length : 0}
                     </AdminTypo.H6>
-                  </HStack>
-                  <Stack>
-                    <AdminTypo.H6 my="2" color="white">
-                      {t("ATTENDANCE_CAMERA_SUBTITLE")}
-                    </AdminTypo.H6>
-                  </Stack>
+                  </HStack> */}
                 </VStack>
               }
-              footerComponent={
-                <HStack space={3} width="100%" justifyContent="space-between">
-                  {error && (
-                    <AdminTypo.H4 style={{ color: "red" }}>
-                      {error}
-                    </AdminTypo.H4>
-                  )}
-                  <AdminTypo.Secondarybutton
-                    shadow="BlueOutlineShadow"
-                    onPress={() => {
-                      updateUserData();
-                      cameraFile
-                        ? setUserData()
-                        : setError("Capture Picture First");
-                      setcameraFile("");
-                      setCameraUrl();
-                    }}
-                  >
-                    {t("FINISH")}
-                  </AdminTypo.Secondarybutton>
-                  <AdminTypo.Secondarybutton
-                    isDisabled={userData?.index + 1 === groupUsers.length}
-                    variant="secondary"
-                    ml="4"
-                    px="5"
-                    onPress={() => {
-                      cameraFile
-                        ? uploadAttendencePicture()
-                        : setError("Capture Picture First");
-                    }}
-                  >
-                    {t("NEXT")}
-                  </AdminTypo.Secondarybutton>
-                </HStack>
-              }
+              // footerComponent={
+              //   <HStack space={3} width="100%" justifyContent="space-between">
+              //     {error && (
+              //       <AdminTypo.H4 style={{ color: "red" }}>
+              //         {error}
+              //       </AdminTypo.H4>
+              //     )}
+              //     <AdminTypo.Secondarybutton
+              //       shadow="BlueOutlineShadow"
+              //       onPress={() => uploadAttendence(userData, PRESENT, true)}
+              //     >
+              //       {t("FINISH")}
+              //     </AdminTypo.Secondarybutton>
+              //     <AdminTypo.Secondarybutton
+              //       isDisabled={userData?.index + 1 === groupUsers.length}
+              //       variant="secondary"
+              //       ml="4"
+              //       px="5"
+              //       onPress={() => uploadAttendence(userData)}
+              //     >
+              //       {t("NEXT")}
+              //     </AdminTypo.Secondarybutton>
+              //   </HStack>
+              // }
               {...{
-                cameraModal,
+                cameraModal: true,
                 setCameraModal: async (item) => {
                   setUserData();
-                  setCameraModal(item);
                 },
                 cameraUrl,
+                onFinish: (e) => uploadAttendence(userData, PRESENT, true),
                 setCameraUrl: async (url, file) => {
                   if (file) {
                     setError("");
                     let formData = new FormData();
+                    formData.append("user_id", userData?.id);
+                    formData.append("document_type", "camp_attendance");
                     formData.append("file", file);
-                    const uploadDoc = await uploadRegistryService.uploadPicture(
+                    const uploadDoc = await uploadRegistryService.uploadFile(
                       formData
                     );
                     if (uploadDoc) {
-                      setcameraFile(uploadDoc);
+                      setCameraFile(uploadDoc);
                     }
                     setCameraUrl({ url, file });
                   } else {
@@ -205,105 +222,102 @@ export default function ConsentForm() {
         _box: { bg: "white" },
       }}
     >
-      <Box py={6} px={4} mb={5}>
-        <AdminTypo.H3 color={"textMaroonColor.400"}>
-          {t("ATTENDANCE")}
-        </AdminTypo.H3>
-
-        {groupUsers?.map((item) => {
-          const document = consents?.find((e) => e.user?.id === item?.id);
-          return (
-            <HStack
-              key={item}
-              bg="white"
-              p="2"
-              my={2}
-              shadow="FooterShadow"
-              rounded="sm"
-              space="1"
-              alignItems={"center"}
-              justifyContent={"space-between"}
-            >
-              <HStack justifyContent="space-between" width={"40%"} space={4}>
-                <HStack alignItems="Center" flex="5">
-                  {item?.profile_photo_1?.id ? (
-                    <ImageView
-                      source={{
-                        uri: item?.profile_photo_1?.name,
-                      }}
-                      // alt="Alternate Text"
-                      width={"45px"}
-                      height={"45px"}
-                    />
-                  ) : (
-                    <IconByName
-                      isDisabled
-                      name="AccountCircleLineIcon"
-                      color="gray.300"
-                      _icon={{ size: "51px" }}
-                    />
-                  )}
-
-                  <VStack
-                    pl="2"
-                    flex="1"
-                    wordWrap="break-word"
-                    whiteSpace="nowrap"
-                    overflow="hidden"
-                    textOverflow="ellipsis"
-                  >
-                    <FrontEndTypo.H3 bold color="textGreyColor.800">
-                      {item?.program_beneficiaries[0]?.enrollment_first_name}
-                      {item?.program_beneficiaries[0]?.enrollment_middle_name &&
-                        ` ${item?.program_beneficiaries[0]?.enrollment_middle_name}`}
-                      {item?.program_beneficiaries[0]?.enrollment_last_name &&
-                        ` ${item?.program_beneficiaries[0]?.enrollment_last_name}`}
-                    </FrontEndTypo.H3>
-                  </VStack>
-                </HStack>
-              </HStack>
-
-              <HStack alignItems={"center"} justifyContent={"space-evenly"}>
-                {document?.status ? (
-                  document?.status === "present" ? (
-                    <HStack space={2}>
-                      <Avatar bg="green.300" size={["15px", "30px"]} />
-                      <FrontEndTypo.H2>{t("PRESENT")}</FrontEndTypo.H2>
+      <GeoLocation
+        getLocation={(lat, long, error) => {
+          if (error) {
+            setError(error);
+          } else {
+            setData({ ...data, lat: `${lat}`, long: `${long}` });
+          }
+        }}
+      />
+      <VStack py={6} px={4} space="6">
+        <HStack justifyContent={"space-between"}>
+          <AdminTypo.H3 color={"textMaroonColor.400"}>
+            {t("LEARNERS")}
+          </AdminTypo.H3>
+          <AdminTypo.H3>({groupUsers?.length || 0})</AdminTypo.H3>
+        </HStack>
+        {/* <FrontEndTypo.Primarybutton onPress={(e) => setUserData(groupUsers[0])}>
+          {t("MARK_ATTENDANCE")}
+        </FrontEndTypo.Primarybutton> */}
+        <VStack space="4">
+          {groupUsers?.map((item) => {
+            return (
+              <HStack key={item} flex="1">
+                <UserCard
+                  _hstack={{ p: 0, space: 1, flex: 1 }}
+                  _vstack={{ py: 2 }}
+                  _image={{ size: 45 }}
+                  leftElement={
+                    (!item?.attendance?.status ||
+                      item?.attendance?.status === PRESENT) && (
+                      <IconByName
+                        onPress={(e) => {
+                          uploadAttendence(item, ABSENT, true);
+                        }}
+                        height="100%"
+                        roundedRight="0"
+                        bg="red.100"
+                        name="CloseCircleLineIcon"
+                        _icon={{ size: "25px", color: "gray" }}
+                      />
+                    )
+                  }
+                  rightElement={
+                    (!item?.attendance?.status ||
+                      item?.attendance?.status === ABSENT) && (
+                      <IconByName
+                        onPress={(e) => {
+                          setUserData(item);
+                        }}
+                        height="100%"
+                        roundedLeft="0"
+                        bg="green.100"
+                        name="CheckboxCircleLineIcon"
+                        _icon={{ size: "25px", color: "gray" }}
+                      />
+                    )
+                  }
+                  title={[
+                    item?.program_beneficiaries[0]?.enrollment_first_name,
+                    item?.program_beneficiaries[0]?.enrollment_middle_name,
+                    item?.program_beneficiaries[0]?.enrollment_last_name,
+                  ]
+                    .filter((e) => e)
+                    .join(" ")}
+                  subTitle={
+                    <HStack>
+                      <RenderAttendee row={item?.attendance || {}} t={t} />
                     </HStack>
-                  ) : (
-                    <HStack space={2}>
-                      <Avatar bg="amber.300" size={["15px", "30px"]} />
-                      <FrontEndTypo.H2>{t("ABSENT")}</FrontEndTypo.H2>
-                    </HStack>
-                  )
-                ) : (
-                  <HStack alignItems={"center"} space={2}>
-                    <Avatar bg="gray.300" size={["15px", "30px"]} />
-                    <FrontEndTypo.H2>{t("PENDING")}</FrontEndTypo.H2>
-                  </HStack>
-                )}
+                  }
+                  image={
+                    item?.profile_photo_1?.fileUrl
+                      ? { urlObject: item?.profile_photo_1 }
+                      : null
+                  }
+                />
               </HStack>
-              <HStack width={"10%"}>
-                {document?.status !== "present" && (
-                  <Pressable
-                    onPress={(e) => {
-                      setUserData(item);
-                      setCameraModal(true);
-                    }}
-                  >
-                    <IconByName
-                      isDisabled
-                      name="CameraLineIcon"
-                      color="blueText.450"
-                      _icon={{ size: "25px" }}
-                    />
-                  </Pressable>
-                )}
-              </HStack>
-            </HStack>
-          );
-        })}
-      </Box>
+            );
+          })}
+        </VStack>
+      </VStack>
     </Layout>
   );
 }
+
+const RenderAttendee = ({ row, t }) => (
+  <Chip
+    py="1px"
+    label={
+      <FrontEndTypo.H5 bold>
+        {row?.fa_is_processed === null
+          ? "-"
+          : row?.fa_is_processed === true
+          ? t("YES") + " " + row?.fa_similarity_percentage?.toFixed(2) + "%"
+          : t("NO")}
+      </FrontEndTypo.H5>
+    }
+    rounded="lg"
+  />
+);
