+++ lib/common-lib/src/components/helper.js
@@ -118,7 +118,8 @@ export const generateUUID = () => {
 }
 
 export const overrideColorTheme = (colorObject = {}, theme = 'joyfull') => {
-  return { ...BaseTheme.colorTheme, ...colorObject }
+  const data = extendTheme(BaseTheme)
+  return { ...data.colors, ...colorObject }
 }
 
 export const DEFAULT_THEME = async (theme) => {
